Squashed commit of the following:
commit 51c1c978163834e3a1b5fb171cb9cbb44d21783b
Author: Jason San Jose <jasonsj@adobe.com>
Date:   Fri Jun 1 13:43:10 2012 -0700

    Implement keyBindingAdded and keyBindingRemoved events

commit f937b2228cc133d578891981db82699cb2a64ea0
Author: Jason San Jose <jasonsj@adobe.com>
Date:   Fri Jun 1 11:17:14 2012 -0700

    Remove keymap

commit 1af5f5887909f682b7da032522f4c93b291becfb
Author: Jason San Jose <jasonsj@adobe.com>
Date:   Thu May 31 16:32:34 2012 -0700

    refactor for  #964

commit 799ca1ef912c1351fbb074485be9c226c1572f80
Author: Jason San Jose <jasonsj@adobe.com>
Date:   Thu May 31 12:31:19 2012 -0700

    refactor key descriptor formatting. remove unused dependency.

commit b2c253ed59e4cdb36d6f80fbb92369a57519ed0d
Author: Jason San Jose <jasonsj@adobe.com>
Date:   Thu May 31 11:44:30 2012 -0700

    Supporrt keyboard shortcut display override
