+++ src/brackets.js
@@ -57,7 +57,6 @@ define(function (require, exports, module) {
         FileViewController      = require("project/FileViewController"),
         FileSyncManager         = require("project/FileSyncManager"),
         KeyBindingManager       = require("command/KeyBindingManager"),
-        KeyMap                  = require("command/KeyMap"),
         Commands                = require("command/Commands"),
         CommandManager          = require("command/CommandManager"),
         BuildInfoUtils          = require("utils/BuildInfoUtils"),
@@ -203,18 +202,6 @@ define(function (require, exports, module) {
             Dialogs.showModalDialog(Dialogs.DIALOG_ID_ABOUT);
         });
     }
-
-    function initKeyHandler() {
-        window.document.body.addEventListener(
-            "keydown",
-            function (event) {
-                if (KeyBindingManager.handleKey(KeyMap.translateKeyboardEvent(event))) {
-                    event.stopPropagation();
-                }
-            },
-            true
-        );
-    }
     
     function _initWindowListeners() {
         // TODO: (issue 269) to support IE, need to listen to document instead (and even then it may not work when focus is in an input field?)
@@ -245,7 +232,6 @@ define(function (require, exports, module) {
 
         _initDragAndDropListeners();
         _initCommandHandlers();
-        initKeyHandler();
         Menus.init(); // key bindings should be initialized first
         _initWindowListeners();
