+++ src/command/KeyBindingManager.js
@@ -22,8 +22,8 @@
  */
 
 
-/*jslint vars: true, plusplus: true, devel: true, nomen: true, indent: 4, maxerr: 50 */
-/*global define, $, brackets */
+/*jslint vars: true, plusplus: true, devel: true, nomen: true, indent: 4, maxerr: 50, regexp: true */
+/*global define, $, brackets, window */
 
 /**
  * Manages the mapping of keyboard inputs to commands.
@@ -31,14 +31,17 @@
 define(function (require, exports, module) {
     'use strict';
 
-    var CommandManager = require("command/CommandManager"),
-        KeyMap = require("command/KeyMap");
+    var CommandManager = require("command/CommandManager");
 
     /**
-     * The currently installed keymap.
-     * @type {KeyMap}
+     * @type {Object}
      */
-    var _keymap = null;
+    var _keyMap = {};
+
+    /**
+     * @type {Object}
+     */
+    var _commandMap = {};
 
     /**
      * Allow clients to toggle key binding
@@ -49,15 +52,239 @@ define(function (require, exports, module) {
      * @private
      */
     function _reset() {
-        _keymap = null;
+        _keymap = {};
+        _commandMap = {}
     }
 
     /**
      * @private
      * Initialize an empty keymap as the current keymap. It overwrites the current keymap if there is one.
+     * builds the keyDescriptor string from the given parts
+     * @param {boolean} hasCtrl Is Ctrl key enabled
+     * @param {boolean} hasAlt Is Alt key enabled
+     * @param {boolean} hasShift Is Shift key enabled
+     * @param {string} key The key that's pressed
+     * @return {string} The normalized key descriptor
      */
-    function _initializeKeymap() {
-        _keymap = KeyMap.create({ "bindings": [], "platform": brackets.platform });
+    function _buildKeyDescriptor(hasCtrl, hasAlt, hasShift, key) {
+        if (!key) {
+            console.log("KeyMap _buildKeyDescriptor() - No key provided!");
+            return "";
+        }
+        
+        var keyDescriptor = [];
+       
+        if (hasAlt) {
+            keyDescriptor.push("Alt");
+        }
+        if (hasShift) {
+            keyDescriptor.push("Shift");
+        }
+
+        if (hasCtrl) {
+            // Windows display Ctrl first, Mac displays Command symbol last
+            if (brackets.platform === "win") {
+                keyDescriptor.unshift("Ctrl");
+            } else {
+                keyDescriptor.push("Ctrl");
+            }
+        }
+
+        keyDescriptor.push(key);
+        
+        return keyDescriptor.join("-");
+    }
+    
+    function _isModifier(left, right, previouslyFound, origDescriptor) {
+        if (!left || !right) {
+            return false;
+        }
+        left = left.trim().toLowerCase();
+        right = right.trim().toLowerCase();
+        var matched = (left.length > 0 && left === right);
+        if (matched && previouslyFound) {
+            console.log("KeyMap normalizeKeyDescriptorString() - Modifier defined twice: " + origDescriptor);
+        }
+        return matched;
+    }
+    
+    /**
+     * normalizes the incoming key descriptor so the modifier keys are always specified in the correct order
+     * @param {string} The string for a key descriptor, can be in any order, the result will be Ctrl-Alt-Shift-<Key>
+     * @return {string} The normalized key descriptor
+     */
+    function normalizeKeyDescriptorString(origDescriptor) {
+        var hasCtrl = false,
+            hasAlt = false,
+            hasShift = false,
+            key = "";
+        
+        origDescriptor.split("-").forEach(function parseDescriptor(ele, i, arr) {
+            if (_isModifier("ctrl", ele, hasCtrl)) {
+                hasCtrl = true;
+            } else if (_isModifier("cmd", ele, hasCtrl, origDescriptor)) {
+                console.log("KeyMap normalizeKeyDescriptorString() - Cmd getting mapped to Ctrl from: " + origDescriptor);
+                hasCtrl = true;
+            } else if (_isModifier("alt", ele, hasAlt, origDescriptor)) {
+                hasAlt = true;
+            } else if (_isModifier("opt", ele, hasAlt, origDescriptor)) {
+                console.log("KeyMap normalizeKeyDescriptorString() - Opt getting mapped to Alt from: " + origDescriptor);
+                hasAlt = true;
+            } else if (_isModifier("shift", ele, hasShift, origDescriptor)) {
+                hasShift = true;
+            } else if (key.length > 0) {
+                console.log("KeyMap normalizeKeyDescriptorString() - Multiple keys defined. Using key: " + key + " from: " + origDescriptor);
+            } else {
+                key = ele;
+            }
+        });
+
+        // Check to see if the binding is for "-".
+        if (key === "" && origDescriptor.search(/^.+--$/) !== -1) {
+            key = "-";
+        }
+        
+        return _buildKeyDescriptor(hasCtrl, hasAlt, hasShift, key);
+    }
+    
+    /**
+     * @private
+     * normalizes the incoming map so all the key descriptors are specified the correct way
+     * @param {map} map The string for a key descriptor, can be in any order, the result will be Ctrl-Alt-Shift-<Key>
+     * @return {map} The normalized map
+     */
+    function _normalizeMap(map) {
+        var finalMap = {};
+        Object.keys(map).forEach(function normalizeKey(ele, i, arr) {
+            var val = map[ele];
+            var normalizedKey = normalizeKeyDescriptorString(ele);
+            if (normalizedKey.length === 0) {
+                console.log("KeyMap _normalizeMap() - Rejecting malformed key: " + ele + " (value: " + val + ")");
+            } else if (!val) {
+                console.log("KeyMap _normalizeMap() - Rejecting key for falsy value: " + ele + " (value: " + val + ")");
+            } else if (finalMap[normalizedKey]) {
+                console.log("KeyMap _normalizeMap() - Rejecting key because it was defined twice: " + ele + " (value: " + val + ")");
+            } else {
+                finalMap[normalizedKey] = val;
+            }
+        });
+        return finalMap;
+    }
+    
+    /**
+     * @private
+     * given a list of bindings, goes through and turns them into a map. The list is filtered based 
+     * on platform, so if a binding has no specific platform or needs to match the given platform
+     * @param [{bindings}] bindings A list of binding objects
+     * @param {string} platform The platform to filter on
+     */
+    function _bindingsToMap(bindings, platform) {
+        var map = {};
+        bindings.forEach(function transformToMap(ele, i, arr) {
+            var keys = Object.keys(ele);
+            var platformIndex = keys.indexOf("platform");
+            if (platformIndex > -1) {
+                if (ele.platform !== platform) {
+                    return; //don't add this to the map
+                }
+                keys.splice(platformIndex, 1);
+            }
+            if (keys.length !== 1) {
+                console.log("KeyMap _bindingsToMap() - bindings list has unknown keys: " + bindings);
+                return;
+            }
+            var key = keys[0];
+            map[key] = ele[key];
+        });
+        return map;
+    }
+    
+    
+    /**
+     * @private
+     * Looks for keycodes that have os-inconsistent keys and fixes them.
+     * @param {number} The keycode from the keyboard event.
+     * @param {string} The current best guess at what the key is.
+     * @return {string} If the key is OS-inconsistent, the correct key; otherwise, the original key.
+     **/
+    function _mapKeycodeToKey(keycode, key) {
+        switch (keycode) {
+        case 186:
+            return ";";
+        case 187:
+            return "=";
+        case 188:
+            return ",";
+        case 189:
+            return "-";
+        case 190:
+            return ".";
+        case 191:
+            return "/";
+        case 192:
+            return "`";
+        case 219:
+            return "[";
+        case 220:
+            return "\\";
+        case 221:
+            return "]";
+        case 222:
+            return "'";
+        default:
+            return key;
+        }
+    }
+    
+    /**
+     * Takes a keyboard event and translates it into a key in a key map
+     */
+    function _translateKeyboardEvent(event) {
+        var hasCtrl = (event.metaKey || event.ctrlKey),
+            hasAlt = (event.altKey),
+            hasShift = (event.shiftKey),
+            key = String.fromCharCode(event.keyCode);
+        
+        //From the W3C, if we can get the KeyboardEvent.keyIdentifier then look here
+        //As that will let us use keys like then function keys "F5" for commands. The
+        //full set of values we can use is here
+        //http://www.w3.org/TR/2007/WD-DOM-Level-3-Events-20071221/keyset.html#KeySet-Set
+        var ident = event.keyIdentifier;
+        if (ident) {
+            if (ident.charAt(0) === "U" && ident.charAt(1) === "+") {
+                //This is a unicode code point like "U+002A", get the 002A and use that
+                key = String.fromCharCode(parseInt(ident.substring(2), 16));
+            } else {
+                //This is some non-character key, just use the raw identifier
+                key = ident;
+            }
+        }
+        
+        // Translate some keys to their common names
+        if (key === "\t") { key = "Tab"; }
+        key = _mapKeycodeToKey(event.keyCode, key);
+
+        return _buildKeyDescriptor(hasCtrl, hasAlt, hasShift, key);
+    }
+    
+    /**
+     * Convert normalized key representation to display appropriate for platform.
+     * @param {!string} descriptor Normalized key descriptor.
+     * @return {!string} Display/Operating system appropriate string
+     */
+    function formatKeyDescriptor(descriptor) {
+        var displayStr;
+        
+        if (brackets.platform === "mac") {
+            displayStr = descriptor.replace(/-/g, "");        // remove dashes
+            displayStr = displayStr.replace("Ctrl", "&#8984");  // Ctrl > command symbol
+            displayStr = displayStr.replace("Shift", "&#8679"); // Shift > shift symbol
+            displayStr = displayStr.replace("Alt", "&#8997");   // Alt > option symbol
+        } else {
+            displayStr = descriptor.replace(/-/g, "+");
+        }
+
+        return displayStr;
     }
 
     /**
@@ -66,25 +293,62 @@ define(function (require, exports, module) {
      * @return {boolean} true if the key is already assigned, false otherwise.
      */
     function _isKeyAssigned(key) {
-        return (_keymap && _keymap.map.hasOwnProperty(key));
+        return (_keyMap && (_keyMap[key] !== undefined));
     }
 
     /**
      * @private
      *
      * @param {string} commandID
-     * @param {string} key - a single shortcut.
+     * @param {string} keyBindingRequest - a single shortcut.
+     * @param {?string} platform - undefined indicates all platforms
+     * @return {?{key: string, displayKey:String}} Returns a record for valid key bindings
      */
-    function _addBinding(commandID, key) {
-        var normalizedKey = KeyMap.normalizeKeyDescriptorString(key);
-        if (!normalizedKey) {
-            console.log("Fail to nomalize " + key);
-        } else if (_isKeyAssigned(normalizedKey)) {
-            console.log("Cannot assign " + normalizedKey + " to " + commandID +
-                        ". It is already assigned to " + _keymap.map[normalizedKey]);
-        } else {
-            _keymap.map[normalizedKey] = commandID;
+    function _addBinding(commandID, keyBindingRequest, platform) {
+        var key,
+            keyBinding = null,
+            normalized,
+            normalizedDisplay,
+            targetPlatform = keyBindingRequest.platform || platform || brackets.platform;
+        
+        // skip if this binding doesn't match the current platform
+        if (targetPlatform !== brackets.platform) {
+            return null;
         }
+        
+        key = (keyBindingRequest.key) || keyBindingRequest;
+        normalized = normalizeKeyDescriptorString(key);
+        
+        // skip if the key binding is invalid 
+        if (!normalized) {
+            console.log("Fail to nomalize " + key);
+            return null;
+        }
+        
+        // skip if the key is already assigned
+        if (_isKeyAssigned(normalized)) {
+            console.log("Cannot assign " + normalized + " to " + commandID +
+                        ". It is already assigned to " + _keyMap[normalized]);
+            return null;
+        }
+        
+        // optional display-friendly string (e.g. CMD-+ instead of CMD-=)
+        normalizedDisplay = (keyBindingRequest.displayKey) ? normalizeKeyDescriptorString(keyBindingRequest.displayKey) : normalized;
+        
+        // 1-to-many commandID mapping to key binding
+        if (!_commandMap[commandID]) {
+            _commandMap[commandID] = [];
+        }
+        
+        keyBinding = {key: normalized, displayKey: normalizedDisplay};
+        _commandMap[commandID].push(keyBinding);
+        
+        // 1-to-1 key binding to commandID
+        _keyMap[normalized] = {commandID: commandID, displayKey: normalizedDisplay};
+        
+        $(exports).triggerHandler("keyBindingAdded", [commandID, keyBinding]);
+        
+        return keyBinding;
     }
 
     /**
@@ -92,11 +356,7 @@ define(function (require, exports, module) {
      * @returns {KeyMap}
      */
     function getKeymap() {
-        if (!_keymap) {
-            return null;
-        }
-        
-        return $.extend({}, _keymap.map);
+        return $.extend({}, _keyMap);
     }
 
     /**
@@ -106,13 +366,12 @@ define(function (require, exports, module) {
      * @return {boolean} true if the key was processed, false otherwise
      */
     function handleKey(key) {
-        if (_enabled && _keymap && _keymap.map[key]) {
-            CommandManager.execute(_keymap.map[key]);
+        if (_enabled && _keyMap[key]) {
+            CommandManager.execute(_keyMap[key].commandID);
             return true;
         }
         return false;
     }
-
 
     // TODO (issue #414): Replace this temporary fix with a more robust solution to handle focus and modality
     /**
@@ -136,42 +395,41 @@ define(function (require, exports, module) {
     /**
      * Add one or more key bindings to a particular Command.
      * 
-     * @param {!string} commandID
-     * @param {!(string | Array.<{key: string, platform: string)}>}  keyBindings - a single key binding
+     * @param {string} commandID
+     * @param {?(string | Array.<{key: string, displayKey: string, platform: string)}>}  keyBindings - a single key binding
      *      or an array of keybindings. Example: "Shift-Cmd-F". Mac and Win key equivalents are automatically
-     *      mapped to each other.
+     *      mapped to each other. Use displayKey property to display a different string (e.g. "CMD+" instead of "CMD=").
      * @param {?string} platform - the target OS of the keyBindings either "mac" or "win". If undefined, all platforms will use
      *      the key binding.
-     *
-     * TODO: MenuItems don't yet update the displayed keyboard shortcut if the keybinding is changed
-     * after menu creation
+     * @return {{key: string, displayKey:String}|Array.<{key: string, displayKey:String}>} Returns record(s) for valid key binding(s)
      */
-    function addBinding(commandID, keyBindings, platform) {
-        if (!_keymap) { _initializeKeymap(); }
-
-        if (!commandID || !keyBindings) {
+    function addBinding(commandID, keyBindingRequests, platform) {
+        if ((commandID === null) || (commandID === undefined) || !keyBindingRequests) {
             return;
         }
+        
+        var normalizedBindings = [],
+            targetPlatform,
+            results;
 
-        if ($.isArray(keyBindings)) {
-            var i, key, targetPlatform;
-            for (i = 0; i < keyBindings.length; i++) {
-                if (keyBindings[i].key !== undefined) {
-                    key = keyBindings[i].key;
-                    targetPlatform = keyBindings[i].platform;
-                } else {
-                    key = keyBindings[i];
-                }
+        if ($.isArray(keyBindingRequests)) {
+            var keyBinding;
+            results = [];
+                                            
+            keyBindingRequests.forEach(function (keyBindingRequest) {
+                targetPlatform = keyBindingRequest.platform || platform || brackets.platform;
+                keyBinding = results.push(_addBinding(commandID, keyBindingRequest, targetPlatform));
                 
-                if (_matchPlatform(targetPlatform)) {
-                    _addBinding(commandID, key);
+                if (keyBinding) {
+                    results.push(keyBinding);
                 }
-            }
-        } else if (_matchPlatform(platform)) {
-            _addBinding(commandID, keyBindings);
+            });
+        } else {
+            targetPlatform = platform || brackets.platform;
+            results = _addBinding(commandID, keyBindingRequests, targetPlatform);
         }
-
-        // TODO: dispatch add event
+        
+        return results;
     }
 
     /**
@@ -181,20 +439,44 @@ define(function (require, exports, module) {
      * @param {string} platform - the intended OS of the key.
      */
     function removeBinding(key, platform) {
-        if (!key || !_keymap || (platform && platform !== brackets.platform)) {
+        if (!key || ((platform !== null) && (platform !== undefined) && (platform !== brackets.platform))) {
             return;
         }
 
-        var normalizedKey = KeyMap.normalizeKeyDescriptorString(key);
+        var normalizedKey = normalizeKeyDescriptorString(key);
+        
         if (!normalizedKey) {
             console.log("Fail to nomalize " + key);
         } else if (_isKeyAssigned(normalizedKey)) {
-            delete _keymap.map[normalizedKey];
-        }
+            var binding = _keyMap[normalizedKey];
+            delete _keyMap[normalizedKey];
 
-        // TODO: dispatch remove event
+            $(exports).triggerHandler("keyBindingRemoved", [binding.commandID, {key: normalizedKey, displayKey: binding.displayKey}]);
+        }
     }
     
+    /**
+     * Retrieve key bindings currently associated with a command
+     *
+     * @param {!string} key - a key-description string that may or may not be normalized.
+     * @return {!Array} An array of associated key bindings.
+     */
+    function getKeyBindings(command) {
+        var bindings = _commandMap[command];
+        return bindings || [];
+    }
+
+    // init
+    window.document.body.addEventListener(
+        "keydown",
+        function (event) {
+            if (handleKey(_translateKeyboardEvent(event))) {
+                event.stopPropagation();
+            }
+        },
+        true
+    );
+
     // unit test only
     exports._reset = _reset;
 
@@ -204,4 +486,6 @@ define(function (require, exports, module) {
     exports.setEnabled = setEnabled;
     exports.addBinding = addBinding;
     exports.removeBinding = removeBinding;
+    exports.formatKeyDescriptor = formatKeyDescriptor;
+    exports.getKeyBindings = getKeyBindings;
 });
